consolidate database_reflection
