+++ src/tests/test_database_reflection.py
@@ -4,8 +4,7 @@ from sqlalchemy.types import VARCHAR
 from testing.postgresql import Postgresql
 from unittest import TestCase
 
-from triage.component.architect import database_reflection as dbreflect
-from triage.database_reflection import table_has_duplicates
+import triage.database_reflection as dbreflect
 
 
 class TestDatabaseReflection(TestCase):
@@ -43,13 +42,13 @@ class TestDatabaseReflection(TestCase):
 
     def test_table_has_duplicates(self):
         self.engine.execute("create table events (col1 int, col2 int)")
-        assert not table_has_duplicates("events", ['col1', 'col2'], self.engine)
+        assert not dbreflect.table_has_duplicates("events", ['col1', 'col2'], self.engine)
         self.engine.execute("insert into events values (1,2)")
         self.engine.execute("insert into events values (1,3)")
-        assert table_has_duplicates("events", ['col1'], self.engine)
-        assert not table_has_duplicates("events", ['col1', 'col2'], self.engine)
+        assert dbreflect.table_has_duplicates("events", ['col1'], self.engine)
+        assert not dbreflect.table_has_duplicates("events", ['col1', 'col2'], self.engine)
         self.engine.execute("insert into events values (1,2)")
-        assert table_has_duplicates("events", ['col1', 'col2'], self.engine)
+        assert dbreflect.table_has_duplicates("events", ['col1', 'col2'], self.engine)
 
     def test_table_has_column(self):
         self.engine.execute("create table incidents (col1 varchar)")
