+++ src/triage/component/architect/entity_date_table_generators.py
@@ -1,7 +1,6 @@
 import verboselogs
 
-from triage.component.architect.database_reflection import table_has_data
-from triage.database_reflection import table_row_count, table_exists, table_has_duplicates
+from triage.database_reflection import table_has_data, table_row_count, table_exists, table_has_duplicates
 
 
 logger = verboselogs.VerboseLogger(__name__)
